refactor: move files/add more accurate types internally
